sand works but liquids just sand rn
