+++ main.py
@@ -1,264 +1,153 @@
+import glfw
+import compushady
+from compushady import HEAP_UPLOAD, Buffer, Swapchain, Texture2D, HEAP_READBACK
+from compushady.formats import R32G32B32A32_FLOAT, R8_UINT, R32_FLOAT, B8G8R8A8_UNORM, R32_UINT, R8G8B8A8_UNORM
+from compushady.shaders import hlsl
+import platform
+import random
+import struct
+import math
 import numpy as np
 import time
-import curses
-import random
 
-size = (32,32)
-clamp = np.ones(size,dtype="uint8")
-world = np.zeros(size,dtype="uint8")
+glfw.init()
+# we do not want implicit OpenGL!
+glfw.window_hint(glfw.CLIENT_API, glfw.NO_API)
 
-# density sorted
-# air
-# smoke
-# fire
-# spark
-# water
-# sand
-# wood
-# rock
-# ===== Render ======
+# use 16 to make d3d11 happy...
+config = compushady.Buffer(16, compushady.HEAP_UPLOAD)
+config_fast = compushady.Buffer(config.size)
 
-col_map = [
-        (600, 600, 1000),
-        (300, 300, 300),
-        (1000, 200, 0),
-        (900, 400, 100),
-        (0, 200, 1000),
-        (1000, 1000, 100),
-        (1000, 1000, 1000),
-        (700,400,200),
-        (20, 40, 60),
+# srv is read only, uav is read & write
+# 0 is air
+# 1 is gas
+# 2 is sand
+# 3 is liquid
+# 4 is solid
+
+mats = [
+    (0.0, [255, 0, 0, 255], 0, "air"),
+    (1.0, [0, 255, 0, 255], 2, "sand"),
+    (0.5, [0, 0, 255, 255], 3, "water"),
 ]
+WIDTH = 1920
+HEIGHT = 1080
+NUM_MATS = len(mats)
 
-def render(scr):
-    for y in range(size[1]):
-        for x in range(size[0]):
-            c = (world[x,size[1]-y-1]+1)//2
-            # erase makes this unnecesary
-            #if c == 0:
-            #    # air hax to get less flickers
-            #    scr.addstr("  ",curses.color_pair(1+c))
-            #else:
-            scr.addstr("██",curses.color_pair(1+c))
-        scr.addstr("\n")
-    do_fps(scr)
+target = Texture2D(WIDTH, HEIGHT, B8G8R8A8_UNORM)
 
-def colour_init():
-    for k,col in enumerate(col_map):
-        curses.init_color(k+1,col[0],col[1],col[2])
-        curses.init_pair(k+1,k+1,k+1)
+# least dry code of all time - like 20 lines of garbage
 
-frame_times = [1 for _ in range(15)]
-current_frame = 0
-last_time = time.time()
-def do_fps(scr):
-    global last_time, current_frame, frame_times
-    current_time = time.time()
-    frame_times[current_frame] = current_time - last_time
-    last_time = current_time
-    current_frame += 1
-    current_frame %= 15
-    fps = 15 / sum(frame_times)
-    scr.addstr(size[1],0,f"FPS: {fps:.0f}",255)
+density = [mat[0] for mat in mats]
+colour = [mat[1] for mat in mats]
+types = [mat[2] for mat in mats]
 
-# ===== Sim =====
+density_buf = compushady.Texture1D(NUM_MATS, R32_FLOAT)
+colour_buf = compushady.Texture1D(NUM_MATS, R8G8B8A8_UNORM)
+types_buf = compushady.Texture1D(NUM_MATS, R8_UINT)
+print(density_buf.size)
+print(colour_buf.size)
+print(types_buf.size)
+staging_buffer_density = Buffer(density_buf.size, HEAP_UPLOAD)
+staging_buffer_colour = Buffer(colour_buf.size, HEAP_UPLOAD)
+staging_buffer_types = Buffer(types_buf.size, HEAP_UPLOAD)
 
-# gas
-# liquid
-# sand
-# solid
-# %2 => ticked
+world = [[random.choice([0, 1, 2]) for y in range(HEIGHT)] for x in range(WIDTH)]
+world_buf = compushady.Texture2D(WIDTH, HEIGHT, R32_UINT)
 
-# hashing is prob slower
-mat_list = [("smoke",0),("fire",0),("spark",1),("water",1),("sand",2),("glass",3),("wood",3),("rock",3)]
 
-def get_mat(name):
-    if name == "air":
-        return 0
-    for k,v in enumerate(mat_list):
-        if v[0] == name:
-            return (k+1)*2
+def copy_bufs():
+    staging_buffer_density.upload(np.array(density, dtype=np.float32))
+    staging_buffer_density.copy_to(density_buf)
+    staging_buffer_colour.upload(np.array(colour, dtype=np.uint8))
+    staging_buffer_colour.copy_to(colour_buf)
+    print(types)
+    staging_buffer_types.upload(np.array(types, dtype=np.uint8))
+    staging_buffer_types.copy_to(types_buf)
 
-def get_type(mid):
-    return mat_list[mid-1][1]
+    staging_buffer_world = Buffer(world_buf.size, HEAP_UPLOAD)
+    staging_buffer_world.upload(np.array(world, dtype=np.uint32))
+    staging_buffer_world.copy_to(world_buf)
 
-def get_name(mid):
-    if mid == 0:
-        return "air"
-    return mat_list[mid-1][0]
+    buffer = Buffer(types_buf.size, HEAP_READBACK)
+    types_buf.copy_to(buffer)
+    read = buffer.readback()
+    stringy = read.hex()
+    print(stringy)
+    # buffer = Buffer(world_buf.size, HEAP_READBACK)
+    # world_buf.copy_to(buffer)
+    # read = buffer.readback()
+    # stringy = read.hex()
+    # print(stringy)
 
-def tick():
-    global world
-    for y in range(size[1]):
-        for x in range(size[0]):
-            material = world[x,y]
-            if material % 2 == 1 or material == 0:
-                continue
-            mid = world[x,y]//2
-            mid = do_reaction(x,y,mid)
-            func_map[get_type(mid)](x,y,mid*2)
 
-def tick_gas(x,y,mid):
-    # unticked pixel
-    if y == size[1]:
-        # top of world
-        return
-    if y < size[1] - 1 and world[x,y+1] < mid:
-        # empty space
-        swap = world[x,y+1]
-        world[x,y+1] = mid + 1
-        world[x,y] = swap
-    else:
-        # filled space
-        direction = random.choice([-1,1])
-        nx = x + direction
-        if nx < 0 or nx >= size[0]:
-            return
-        if world[nx,y] < mid:
-            # empty space next to
-            swap = world[nx,y]
-            world[nx,y] = mid + 1
-            world[x,y] = swap
- 
-def tick_liquid(x,y,mid): 
-    # unticked pixel
-    if world[x,y-1] < mid and y != 0:
-        # empty space
-        swap = world[x,y-1]
-        world[x,y-1] = mid + 1
-        world[x,y] = swap
-    else:
-        # filled space
-        direction = random.choice([-1,1])
-        nx = x + direction
-        if nx < 0 or nx >= size[0]:
-            return
-        if world[nx,y] < mid:
-            # empty space next to
-            swap = world[nx,y]
-            world[nx,y] = mid + 1
-            world[x,y] = swap
+copy_bufs()
 
-def tick_sand(x,y,mid):
-    if y == 0:
-        # bottom of world
-        return
-    below = world[x,y-1]
-    if below < mid:
-        pass
-        # empty space
-        world[x,y-1] = mid + 1
-        world[x,y] = below
-    else:
-        # filled space
-        direction = random.choice([-1,1])
-        nx = x + direction
-        if nx < 0 or nx >= size[0]:
-            return
-        below_side = world[nx,y-1]
-        if below_side < mid:
-            # empty space next to
-            world[nx,y] = mid + 1
-            world[x,y] = below_side
+with open("compute.hlsl") as f:
+    shader_compute = hlsl.compile(
+        f
+        .read()
+        .replace("1/*$WIDTH*/", str(WIDTH))
+        .replace("1/*$HEIGHT*/", str(HEIGHT))
+        .replace("1/*$NUM_MATS*/", str(NUM_MATS))
+    )
+compute = compushady.Compute(shader_compute, cbv=[config_fast], srv=[
+                             types_buf,density_buf], uav=[world_buf])
 
-def tick_solid(x,y,mid):
-    pass # solids don't do anything
+with open("render.hlsl") as f:
+    shader_render = hlsl.compile(
+        f
+        .read()
+        .replace("1/*$WIDTH*/", str(WIDTH))
+        .replace("1/*$HEIGHT*/", str(HEIGHT))
+        .replace("1/*$NUM_MATS*/", str(NUM_MATS))
+    )
+# , srv=[world_buf, colour_buf]
+render = compushady.Compute(
+    shader_render, srv=[world_buf, colour_buf], uav=[target])
 
-def tick_rock(x,y,mid):
-    pass # rocks dont do much
+window = glfw.create_window(
+    target.width, target.height, 'Random', None, None)
 
-func_map = [tick_gas, tick_liquid, tick_sand, tick_solid]
+if platform.system() == 'Windows':
+    swapchain = compushady.Swapchain(glfw.get_win32_window(
+        window), compushady.formats.B8G8R8A8_UNORM, 2)
+elif platform.system() == 'Darwin':
+    # macos
+    from compushady.backends.metal import create_metal_layer
+    ca_metal_layer = create_metal_layer(glfw.get_cocoa_window(
+        window), compushady.formats.B8G8R8A8_UNORM)
+    swapchain = compushady.Swapchain(
+        ca_metal_layer, compushady.formats.B8G8R8A8_UNORM, 2)
+else:
+    swapchain = compushady.Swapchain((glfw.get_x11_display(), glfw.get_x11_window(
+        window)), compushady.formats.B8G8R8A8_UNORM, 2)
 
-reactions = [
-        ("fire","air",0.02),
-        ("fire","smoke",0.1),
-        ("fire","spark",0.02),
-        ("fire+wood","spark+fire",0.95),
-        ("spark+wood","fire+fire",0.8),
-        ("spark","fire",0.3),
-        ("smoke","air",0.02),
-        ("fire+smoke","smoke+smoke",0.4),
-        ("fire+sand","fire+glass",0.05),
-        ("fire+air","air+fire",0.4),
-]
+count = 0
+start = None
+multiplier = 0
+while not glfw.window_should_close(window):
+    glfw.poll_events()
 
-def do_reaction(x,y,mid):
-    nx = x + random.choice([-1,0,1])
-    ny = y + random.choice([-1,0,1])
-    if x != nx or y != ny:
-        if nx >= 0 and nx < size[0] and ny >= 0 and ny < size[1]:
-            new = get_reaction(get_name(world[x,y]//2),get_name(world[nx,ny]//2))
-            world[x,y] = get_mat(new[0][0])
-            world[nx,ny] = get_mat(new[0][1])
-            return world[x,y]//2
-    return mid
+    # update "push constants*/" or whatever compushady calls them
+    # config.upload(struct.pack('f', abs(math.sin(multiplier))))
+    # config.copy_to(config_fast)
+    render.dispatch(target.width // 8, target.height // 8, 1)
+    for x in range(3):
+        for y in range(3): # prevent weird physics
+            config.upload(np.array([x,y,random.random()*16],dtype=np.uint32))
+            config.copy_to(config_fast)
+            # print(compute)
+            # print(shader_compute)
+            compute.dispatch(target.width // 8, target.height // 8, 1)
+    swapchain.present(target)
+    # time.sleep(0.5)
+    if start is None:
+        start = time.time()
+    # multiplier += 0.02
+    count += 1
+print(count/(time.time()-start))
 
-def get_reaction(a,b):
-    for reaction in reactions:
-        data = match_reaction(reaction,a,b)
-        if data[1]:
-            return data
-    return ((a,b),False)
+swapchain = None  # this ensures the swapchain is destroyed before the window
 
-def match_reaction(reaction,a,b):
-    parts = reaction[0].split("+")
-    parts_other = reaction[1].split("+")
-    if len(parts) == 1:
-        if a == parts[0]:
-            if random.random() < reaction[2]:
-                return ((parts_other[0],b),True)
-    else:
-        if a == parts[0] and b == parts[1]:
-            if random.random() < reaction[2]:
-                return ((parts_other[0],parts_other[1]),True)
-    return ((a,b),False)
-
-def clean():
-    global world
-    for y in range(size[1]):
-        for x in range(size[0]):
-            if world[x,y] == 0:
-                continue
-            world[x,y] = ((world[x,y]) // 2) * 2
-
-def world_init():
-    for y in range(size[1]):
-        for x in range(size[0]):
-            ty = get_mat("air")
-            if random.random() < 0.1:
-                ty = get_mat("fire")
-            elif random.random() < 0.1:
-                ty = get_mat("sand")
-            elif y * 0.1 < random.random():
-                ty = get_mat("water")
-            elif random.random() < y * 0.05:
-                ty = get_mat("wood")
-            world[x,y] = ty
-
-# ===== Control =====
-
-def init():
-    colour_init()
-    world_init()
-
-def main(scr):
-    if not curses.can_change_color():
-        raise Exception("Doesn't support fancy colours")
-    scr.nodelay(True)
-    init()
-    while True:
-        #scr.clear()
-        scr.erase() # magically fixes all the flickering
-        #time.sleep(0.1)
-        scr.addstr(str())
-        tick()
-        clean()
-        render(scr)
-        scr.refresh()
-        try:
-            key = scr.getkey()
-            scr.addstr(key)
-        except:
-            pass
-
-curses.wrapper(main)
+glfw.terminate()
